Manually populate Filepicker input and preview
- to get around conflict with Turbolinks.
- Because Turbolinks replaces the body content via AJAX, we end up not
reloading the javascript and therefore not populating the Filepicker
widget when following a Turbo-link (as on our Scholarship wizard).
- To fix, use an input field with data-fp attributes for Filepicker,
instead of relying on the provided widget.

Also auto-populate our image preview when a new image is uploaded.
- Manually via Javascript to update the img's src attr.

NB: Filestack-rails gem is still part of the bundle, but not currently
used. Consider removing.